+++ CHANGELOG.md
@@ -10,7 +10,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.2.0] - TBA
+## [1.3.0] - TBA
+# Changed
+- Added reload function which adds shortcuts without restarting the program
+- Added mechanism to update config file and repair broken sections
+- Added better exception handling for issues in config file
+- Added contributions file
+- Moved source files into textscript subdirectory and refactored code
+- Added notifications for added or removed textblocks
+- Reorganized classes so that methods make more sense
+- Renamed internal use methods according to PEP-8
+- Renamed Settings.py as ConfigUtils.py as this is more accurate
+
+## [1.2.0] - 2020-01-23
 ### Added
 - Support for three directories: default, local, and remote
 - Usage statistic printout during program start